Create Deno server, make statusBox communal
Created a server. Clients connect to the server through websockets and the server serves them index.html. This works locally when I deploy from the directory, but I'm unsure of what changes I may need to make to get it to work with Deno Deploy (which I think I set up to deploy from this github repo).

When a client updates statusBox, it messages the server, which then can message the other clients to update their statusBox. Later down the line, this will be expanded to other features but I wanted to get basic functionality down first.

There are still some minor issues with this implementation for text boxes. Namely, because the server tells the client to update the html value of the textbox, their cursor goes to the bottom of the box instead of staying where it was originally.