Update Deno server to use serve instead of oak
As it turns out, the EA version of Deno Deploy that we are using to host the server does not support oak. The code has been rewritten to use Deno serve instead. Any requests to url pathnames other than "/" are assumed to be for other files in the project (js, css, etc), and in case of error, a 404 response is returned