Try to fix Deno Deploy issue with websocket request
For unknown reasons, it seems as though somewhere in the process of the client requesting a websocket, the url is being turned back into http/https from ws/wss, or otherwise not having the "upgrade" header. This is not an issue locally, so I'm not sure why it is for Deno Deploy.

I can't update the code for the Deno server on Deno Deploy without pushing this code to github, so unfortunately I'm forced to commit many changes until I can find/resolve the issue. As such, I'm making a branch (server-bugfix) for this to not affect the main branch