Deno server is working!
Updated the readme to link to the server, cleaned up commented-out code. The Deno server works for sending information to all connected clients (which for now is only the text in the textbox). Later I plan to create a landing page to create/join a "room" so multiple games can be played at once.

The websocket issue turned out to be that Deno Deploy was treating the site as statically deployed from index.html instead of deploying the server itself. I'm not sure if I had to change much in the code itself, just in how Deno launched it.