Added client/server image upload
Added code on the client and server side  so that clients can upload images to be stored on the server and shared with all other clients. Currently, the only element we do image uploads for is the background map (id "mapBg"). The code now supports uploading a new background map image in "mapBg" to be seen by all other clients on the site.