Update server to send and track updatedItems
Previously, server would request this information from another client. This was messy. Now the server stores this information itself so it can update connecting clients quickly. This change also allows the server to broadcast this info when a new server connects--though testing this part will involve deploying on Deno Deploy