Performance improvements
Use image url for image uploads instead of fetching the bytes.

Keep track of updatedItems in Deno KV instead of needing to broadcast them when a new server joins (still needs testing on Deno Deploy to ensure this works for multiple servers).

This logic fix should make it easier to create "rooms" for different games when the time comes since now we can set the updatedItems for a certain room code instead.