Add logo and nav pages
Created a cute logo for the project :)

Added in a header with navigation links (will recolor later). Also created dummy html pages for the other tabs.

Had to minorly update the server code to serve the logo. I changed the code so this will not be an issue for any future static images now, so long as they are added to the "images" folder.