Begin styling
Add custom colors and recolor elements on all the pages. Also add in the room code at the top of the page when you've joined a room.

There's still a lot more styling that needs to be done, next commit I'll probably update the font and fix some of the padding and spacing issues.