Fix padding & layout, add font, add mobile menu
Fix some padding issues on the header and footer. Add the Quicksand font to the website. Fix the layout of the header on desktop and create a mobile menu.

Also renamed combatmap.js to layout.js since I think the only information not shared through the websocket (and therefore the code should be in client.js) is the website layout.

(note: the z-20 class may not be necessary but I'm keeping it on there just in case to make sure the button shows up on top of the header)