Add tab image & animation
The currently-selected tab is "eaten" by the map mimic creature (mm_selector.png). When hovering over a different tab, the creature bounces at the bottom. It's very cute and I'm very proud of it :).

Also updated padding slightly to work better as the window gets narrower and removed an old print statement from the server code.