Add footer creature & textbox, move to bottom
Move the footer to the bottom of the page and add the map mimic creature and a "text bubble" pointing the user to the github repository.

The space-y-4 class was causing a "bounce" effect on the footer on page load due to messing with its top/bottom margins, so I added a margin to the header bottom and padding around the footer instead to get rid of this bounce.