Set up room layout and add buttons
Buttons are currently non-functional other than the one that uploads a new map image, will add functionality to the others in a later commit.

The image fits on screen without needing to scroll, regardless of screen size (will shrink and maintain proportions as need be). Two columns on larger screens (image and initiative box), one on smaller screens. On desktop (and large tablet) the image will also grow to best fit the given area (see below for what happens on mobile).

Tiny issue with smaller screens is that the image doesn't grow, so if you upload a really small image it won't auto-grow on mobile like it does on desktop. This is because I had to set the height class to h-auto when switching to one column layout in order for the buttons to not overlap with the initiative box (due to going over the fixed h-* value by having the first row of the container take up the entire height of the container--and the container must have this fixed height for layout reasons). That being said, you'd have to upload something stupidly tiny for this to happen, and I've spent too long trying to get the layout to behave as-is so I'm not fixing this, at least not right now.